Add files via upload
I tried pushing this file from R this weekend and it appeared to have worked and had uploaded my markdown file to github, but when checking my github again this morning I did not see the file in my repository. Here is the PDF version instead since my files still aren't pushing correctly to github. I apologize for the later submission due to this!